% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function_bootstrap.R
\name{para_boot}
\alias{para_boot}
\title{Final Project Parametric Bootstrap}
\usage{
para_boot(data, example, B)
}
\arguments{
\item{data}{the data set for the example}

\item{example}{the name of example, in our case we focus on "tortoise"}

\item{B}{sample size of bootstrap}
}
\value{
A list with summary statistics including point estimates of coefficients, standard errors, confidence intervals and p-value for hypothesis test.
}
\description{
Final Project Parametric Bootstrap
}
\examples{
##### Tortoise #####
## Load data
library(readr)
setwd("/Users/jinruyi/Xu.Jin.4864final.package/R")
tortoise <- read_csv("gopher_tortoise.csv")

## Choose the desired sample size for bootstrap
B <- 1000

## Apply the para_boot function to the data set with selected B
para_boot(tortoise, "tortoise", B)
}
